+++ lua/csharp/types.lua
@@ -11,21 +11,6 @@
 --- @field FileName string
 --- @field Range OmnisharpRange
 
---- @class OmnisharpMetadataSource
---- @field AssemblyName string
---- @field TypeName string
---- @field ProjectName string
---- @field VersionNumber string
---- @field Language string
-
---- @class OmnisharpMetadataResponse
---- @field Source string
---- @field SourceName string
-
---- @class OmnisharpDefinition
---- @field Location OmnisharpLocation
---- @field MetadataSource OmnisharpMetadataSource?
-
 --- @class OmnisharpTextChange
 --- @field NewText string
 --- @field StartLine number
@@ -54,7 +39,7 @@
 
 --- @class LspHandlerContext<T>: {["params"]: T, ["client_id"]: number, ["bufnr"]: number, ["method"]: string}
 
---- @class LspRequestSyncResponse<T>: {["result"]: T, ["err"] = LspError|nil}
+--- @class LspRequestSyncResponse<T>: {["result"]: T, ["err"]: LspError|nil}
 
 --- @class LspWorkspaceEdit
 --- @field changes table<string, LspTextEdit[]>